feat: Add roamer density-dependent removal
This commit adds a new function `roamer_density_dependent_removal` to the `mortality.rs` module. This function removes roamers from the model when there are more males than females. It calculates the number of adult females and the number of roamers and removes random roamers until the number of roamers is equal to or less than the number of adult females.

The function is called in the `main` function on the first day of each year. This ensures that roamers are removed when there is an imbalance in the sex ratio.

This change improves the realism of the model by simulating the natural behavior of roamers leaving the area when there are more males than females.

```