feat: Add id_to_index mapping in InteractionLayer
This commit adds a new field `id_to_index` to the `InteractionLayer` struct. This field is a `HashMap` that maps the individual ID to its index in the `entities` vector. The `id_to_index` mapping is used to efficiently update and query entities in the KD-tree.

This change improves the performance of entity updates and queries in the interaction layer by providing a fast lookup mechanism.