refactor: Add uuid dependency and generate unique simulation ID
This commit adds the `uuid` dependency to the project by updating the `Cargo.toml` and `Cargo.lock` files. It also introduces a new function `generate_unique_simulation_id` in the `utility.rs` file to generate a unique simulation ID using the UUID library. This enhancement improves the functionality of the code by providing a reliable way to generate unique IDs for simulations.