refactor: Add time_of_infection and infection_stage fields to new group members
This commit adds the `time_of_infection` and `infection_stage` fields to the `new_member` struct in the `reproduction.rs` file. The `time_of_infection` field is set to `None` and the `infection_stage` field is set to `InfectionStage::NotInfected`. This change enhances the data recorded for new group members and prepares for future infection tracking in the simulation.